Added error handling
The WebSocketJSONStream package does not have error handling out of the box, and occasionally errors can occur if the backend needs to send updates to a client when that client is reconnecting. This caused the backend to crash. So, this error handling is required to prevent the backend from crashing.